requirements.txt
flask
python-dotenv
groq
supabase
postgrest
gunicorn
